Bugs

This data frame contains data from an experiment to see if insects are more attracted to some colors than to others. The researchers prepared colored cards with a sticky substance so that insects that landed on them could not escape. The cards were placed in a field of oats in July. Later the researchers returned, collected the cards, and counted the number of cereal leaf beetles trapped on each card.

Format

A data frame with 24 observations on the following 2 variables.

color

color of card; one of B(lue) G(reen) W(hite) Y(ellow)

trapped

number of insects trapped on the card

Source

M. C. Wilson and R. E. Shade, Relative attractiveness of various luminescent colors to the cereal leaf beetle and the meadow spittlebug, Journal of Economic Entomology 60 (1967), 578–580.

Examples

data(Bugs)
#> Warning: data set ‘Bugs’ not found
favstats(trapped ~ color, data = Bugs)
#>   color min    Q1 median    Q3 max     mean       sd n missing
#> 1     B   7 11.75   15.0 19.00  21 14.83333 5.344779 6       0
#> 2     G  15 26.75   34.5 38.50  41 31.50000 9.914636 6       0
#> 3     W  12 13.25   15.5 17.00  21 15.66667 3.326660 6       0
#> 4     Y  38 45.25   46.5 47.75  59 47.16667 6.794606 6       0