Concrete Compressive Strength Data

These data were collected by I-Cheng Yeh to determine how the compressive strength of concrete is related to its ingredients (cement, blast furnace slag, fly ash, water, superplasticizer, coarse aggregate, and fine aggregate) and age.

Format

Concrete is a data frame with the following variables.

limestone

percentage of limestone

water

water-cement ratio

strength

compressive strength (MPa) after 28 days

References

Appeared in Devore's "Probability and Statsistics for Engineers and the Sciences (6th ed). The variables have been renamed.