Family smoking

A cross-tabulation of whether a student smokes and how many of his or her parents smoke from a study conducted in the 1960's.

Format

A data frame with 5375 observations on the following 2 variables.

Student a factor with levels DoesNotSmoke Smokes

Parents a factor with levels NeitherSmokes OneSmokes BothSmoke

Source

S. V. Zagona (ed.), Studies and issues in smoking behavior, University of Arizona Press, 1967.

References

The data also appear in

Brigitte Baldi and David S. Moore, The Practice of Statistics in the Life Sciences, Freeman, 2009.

Examples

data(FamilySmoking)

#> Warning: data set ‘FamilySmoking’ not found

xchisq.test( tally(parents ~ student, data = FamilySmoking) )

#> 
#> 	Pearson's Chi-squared test
#> 
#> data:  x
#> X-squared = 37.566, df = 2, p-value = 6.959e-09
#> 
#>   1168      188  
#> (1102.71) ( 253.29)
#> [ 3.8655] [16.8288]
#> < 1.966> <-4.102>
#>    
#>   1823      416  
#> (1820.78) ( 418.22)
#> [ 0.0027] [ 0.0118]
#> < 0.052> <-0.109>
#>    
#>   1380      400  
#> (1447.51) ( 332.49)
#> [ 3.1488] [13.7086]
#> <-1.774> < 3.703>
#>    
#> key:
#> 	observed
#> 	(expected)
#> 	[contribution to X-squared]
#> 	<Pearson residual>