Pass the Pigs

This data set contains information collected from rolling the pair of pigs (found in the game "Pass the Pigs") 6000 times.

Format

A data frame with 6000 observations on the following 6 variables.

roll

roll number (1-6000)

blackScore

numerical code for position of black pig

black

position of black pig coded as a factor

pinkScore

numerical code for position of pink pig

pink

position of pink pig coded as a factor

score

score of the roll

height

height from which pigs were rolled (5 or 8 inches)

start

starting position of the pigs (0 = both pigs backwards, 1 = one bacwards one forwards, 2 = both forwards)

Source

John C. Kern II, Duquesne University (kern@mathcs.duq.edu)

Details

In "Pass the Pigs", players roll two pig-shaped rubber dice and earn or lose points depending on the configuration of the rolled pigs. Players compete individually to earn 100 points. On each turn, a player rolls he or she decides to stop or until "pigging out" or

The pig configurations and their associated scores are

1 = Dot Up (0)

2 = Dot Down (0)

3 = Trotter (5)

4 = Razorback (5)

5 = Snouter (10)

6 = Leaning Jowler (15)

7 = Pigs are touching one another (-1; lose all points)

One pig Dot Up and one Dot Down ends the turn (a "pig out") and results in 0 points for the turn. If the pigs touch, the turn is ended and all points for the game must be forfeited. Two pigs in the Dot Up or Dot Down configuration score 1 point. Otherwise, The scores of the two pigs in different configurations are added together. The score is doubled if both both pigs have the same configuration, so, for example, two Snouters are worth 40 rather than 20.

Examples

data(Pigs)
#> Warning: data set ‘Pigs’ not found
tally( ~ black, data = Pigs )
#> black
#>         Dot Up       Dot Down        Trotter      Razorback        Snouter 
#>           1796           2063            551           1354            184 
#> Leaning Jowler       Touching 
#>             29             23 
if (require(tidyr)) {
  Pigs %>% 
  select(roll, black, pink) %>%
  gather(pig, state, black, pink) %>%
  tally( state ~ pig, data = ., format = "prop", margins = TRUE)
}
#>                 pig
#> state                  black        pink
#>   Dot Down       0.343833333 0.352833333
#>   Dot Up         0.299333333 0.301833333
#>   Leaning Jowler 0.004833333 0.007333333
#>   Razorback      0.225666667 0.220000000
#>   Snouter        0.030666667 0.029833333
#>   Touching       0.003833333 0.003833333
#>   Trotter        0.091833333 0.084333333
#>   Total          1.000000000 1.000000000