Maze tracing and scents

Subjects were asked to to complete a pencil and paper maze when they were smelling a floral scent and when they were not.

Format

A data frame with 21 observations on the following 12 variables.

id

ID number

sex

a factor with levels F andM

smoker

a factor with levels N, Y

opinion

opinion of the odor (indiff, neg, or pos

)
age

age of subject (in years)

first

which treatment was first, scented or unscented

u1

time (in seconds) in first unscented trial

u2

time (in seconds) in second unscented trial

u3

time (in seconds) in third unscented trial

s1

time (in seconds) in first scented trial

s2

time (in seconds) in second scented trial

s3

time (in seconds) in third scented trial

Source

These data are also available at DASL, the data and story library (https://dasl.datadescription.com/).

References

Hirsch, A. R., and Johnston, L. H. "Odors and Learning," Smell & Taste Treatment and Research Foundation, Chicago.

Examples

data(Scent)
#> Warning: data set ‘Scent’ not found
summary(Scent)
#>        id     sex    smoker   opinion        age              first   
#>  Min.   : 1   F:10   N:13   indiff: 4   Min.   :15.00   scented  :10  
#>  1st Qu.: 6   M:11   Y: 8   neg   : 7   1st Qu.:26.00   unscented:11  
#>  Median :11                 pos   :10   Median :35.00                 
#>  Mean   :11                             Mean   :36.57                 
#>  3rd Qu.:16                             3rd Qu.:43.00                 
#>  Max.   :21                             Max.   :65.00                 
#>        u1              u2              u3              s1       
#>  Min.   :32.80   Min.   :26.80   Min.   :25.70   Min.   :28.30  
#>  1st Qu.:40.90   1st Qu.:38.00   1st Qu.:37.20   1st Qu.:49.60  
#>  Median :49.50   Median :46.80   Median :43.20   Median :53.60  
#>  Mean   :53.92   Mean   :48.94   Mean   :47.18   Mean   :55.53  
#>  3rd Qu.:60.10   3rd Qu.:57.90   3rd Qu.:58.00   3rd Qu.:67.30  
#>  Max.   :93.80   Max.   :91.90   Max.   :77.40   Max.   :77.50  
#>        s2               s3       
#>  Min.   : 25.10   Min.   :24.80  
#>  1st Qu.: 36.80   1st Qu.:34.40  
#>  Median : 44.00   Median :42.90  
#>  Mean   : 48.31   Mean   :43.32  
#>  3rd Qu.: 53.40   3rd Qu.:48.40  
#>  Max.   :126.60   Max.   :64.50