Nonlinear maximization and minimization

nlmin and nlmax are thin wrappers around nlm, a non-linear minimizer. nlmax avoids the necessity of modifying the function to construct a minimization problem from a problem that is naturally a maximization problem. The summary method for the resulting objects provides output that is easier for humnans to read.

Usage

nlmax(f, ...)

nlmin(f, ...)

# S3 method for class 'nlmax'
summary(object, nsmall = 4, ...)

# S3 method for class 'nlmin'
summary(object, nsmall = 4, ...)

Arguments

f

a function to optimize

...

additional arguments passed to nlm. Note that p is a required argument for nlm. See the help for nlm for details.

object

an object returned from nlmin or nlmax

nsmall

a numeric passed through to format

Examples

summary( nlmax( function(x) 5 - 3*x - 5*x^2, p=0 ) )
#> 
#>        Maximum: 5.4500
#>       Estimate:-0.3000005
#>       Gradient:-8.881784e-10
#>     Iterations: 1
#> 
#> Relative gradient is close to zero, current iterate is probably an
#> approximate solution.[Code=1]