Concrete Compressive Strength Data

These data were collected by I-Cheng Yeh to determine how the compressive strength of concrete is related to its ingredients (cement, blast furnace slag, fly ash, water, superplasticizer, coarse aggregate, and fine aggregate) and age.

ConcreteAll

Format

Concrete is a data frame with the following variables.

limestone percentage of limestone

water water-cement ratio

strength compressive strength (MPa) after 28 days

Source

Data were obtained from the Machine Learning Repository (http://archive.ics.uci.edu/ml/) where they were deposited by I-Cheng Yeh (icyeh at chu.edu.tw) who retains the copyright for these data.

References

Appeared in Devore's "Probability and Statsistics for Engineers and the Sciences (6th ed). The variables have been renamed.

#' Concrete Compressive Strength Data

These data were collected by I-Cheng Yeh to determine how the compressive strength of concrete is related to its ingredients (cement, blast furnace slag, fly ash, water, superplasticizer, coarse aggregate, and fine aggregate) and age.

I-Cheng Yeh (1998), "Modeling of strength of high performance concrete using artificial neural networks," Cement and Concrete Research, Vol. 28, No. 12, pp. 1797-1808.

Examples

data(Concrete)