Corn Yield

William Gosset analyzed data from an experiment comparing the yield of regular and kiln-dried corn.

Format

A data frame with 11 observations on the following 2 variables.

reg yield of regular corn (lbs/acre)

kiln yield of kiln-dried corn (lbs/acre)

Source

These data are also available at DASL, the data and story library (https://dasl.datadescription.com/).

Details

Gosset (Student) reported on the results of seeding plots with two different kinds of seed. Each type of seed (regular and kiln-dried) was planted in adjacent plots, accounting for 11 pairs of "split" plots.

References

W.S. Gosset, "The Probable Error of a Mean," Biometrika, 6 (1908), pp 1-25.

Examples

Corn2 <- stack(Corn)
names(Corn2) <- c('yield','treatment')
lm(yield ~ treatment, data = Corn2)
#> 
#> Call:
#> lm(formula = yield ~ treatment, data = Corn2)
#> 
#> Coefficients:
#>   (Intercept)  treatmentkiln  
#>       1841.45          33.73  
#> 
t.test(yield ~ treatment, data = Corn2)
#> 
#> 	Welch Two Sample t-test
#> 
#> data:  yield by treatment
#> t = -0.23413, df = 19.983, p-value = 0.8173
#> alternative hypothesis: true difference in means between group reg and group kiln is not equal to 0
#> 95 percent confidence interval:
#>  -334.2292  266.7746
#> sample estimates:
#>  mean in group reg mean in group kiln 
#>           1841.455           1875.182 
#> 
t.test(Corn$reg, Corn$kiln)
#> 
#> 	Welch Two Sample t-test
#> 
#> data:  Corn$reg and Corn$kiln
#> t = -0.23413, df = 19.983, p-value = 0.8173
#> alternative hypothesis: true difference in means is not equal to 0
#> 95 percent confidence interval:
#>  -334.2292  266.7746
#> sample estimates:
#> mean of x mean of y 
#>  1841.455  1875.182 
#>