FUSION type 2 diabetes study

Phenotype and genotype data from the Finland United States Investigation of NIDDM (type 2) Diabetes (FUSION) study.

Format

Data frames with the following variables.

id subject ID number for matching between data sets

t2d a factor with levels case control

bmi body mass index

sex a factor with levels F M

age age of subject at time phenotypes were colelcted

smoker a factor with levels former never occasional regular

chol total cholesterol

waist waist circumference (cm)

weight weight (kg)

height height (cm)

whr waist hip ratio

sbp systolic blood pressure

dbp diastolic blood pressure

marker RS name of SNP

markerID numeric ID for SNP

allele1 first allele coded as 1 = A, 2 = C, 3 = G, 4 = T

allele2 second allele coded as 1 = A, 2 = C, 3 = G, 4 = T

genotype both alleles coded as a factor

Adose number of A alleles

Cdose number of C alleles

Gdose number of G alleles

Tdose number of T alleles

Source

Similar to the data presented in

Laura J. Scott, Karen L. Mohlke, Lori L. Bonnycastle, Cristen J. Willer, Yun Li, William L. Duren, Michael R. Erdos, Heather M. Stringham, Pe- ter S. Chines, Anne U. Jackson, Ludmila Prokunina-Olsson, Chia-Jen J. Ding, Amy J. Swift, Narisu Narisu, Tianle Hu, Randall Pruim, Rui Xiao, Xiao- Yi Y. Li, Karen N. Conneely, Nancy L. Riebow, Andrew G. Sprau, Maurine Tong, Peggy P. White, Kurt N. Hetrick, Michael W. Barnhart, Craig W. Bark, Janet L. Goldstein, Lee Watkins, Fang Xiang, Jouko Saramies, Thomas A. Buchanan, Richard M. Watanabe, Timo T. Valle, Leena Kinnunen, Goncalo R. Abecasis, Elizabeth W. Pugh, Kimberly F. Doheny, Richard N. Bergman, Jaakko Tuomilehto, Francis S. Collins, and Michael Boehnke, A genome-wide association study of type 2 diabetes in Finns detects multiple susceptibility vari- ants, Science (2007).

Examples

data(Pheno); data(FUSION1); data(FUSION2)
#> Warning: data set ‘Pheno’ not found
#> Warning: data set ‘FUSION1’ not found
#> Warning: data set ‘FUSION2’ not found
FUSION1m <- merge(FUSION1, Pheno, by = "id", all.x = FALSE, all.y = FALSE) 
xtabs( ~ t2d + genotype, data = FUSION1m) 
#>          genotype
#> t2d        GG  GT  TT
#>   case    737 375  48
#>   control 835 309  27
xtabs( ~ t2d + Gdose, data = FUSION1m) 
#>          Gdose
#> t2d         0   1   2
#>   case     48 375 737
#>   control  27 309 835
chisq.test( xtabs( ~ t2d + genotype, data = FUSION1m ) )
#> 
#> 	Pearson's Chi-squared test
#> 
#> data:  xtabs(~t2d + genotype, data = FUSION1m)
#> X-squared = 18.306, df = 2, p-value = 0.0001059
#> 
f1.glm <- glm( factor(t2d) ~ Gdose, data = FUSION1m, family = binomial) 
summary(f1.glm)
#> 
#> Call:
#> glm(formula = factor(t2d) ~ Gdose, family = binomial, data = FUSION1m)
#> 
#> Deviance Residuals: 
#>    Min      1Q  Median      3Q     Max  
#> -1.232  -1.232   1.124   1.124   1.410  
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)    
#> (Intercept) -0.53183    0.13402  -3.968 7.24e-05 ***
#> Gdose        0.32930    0.07737   4.256 2.08e-05 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 3231.4  on 2330  degrees of freedom
#> Residual deviance: 3213.0  on 2329  degrees of freedom
#> AIC: 3217
#> 
#> Number of Fisher Scoring iterations: 3
#>