Goals and popularity factors for school kids

Subjects were students in grades 4-6 from three school districts in Michigan. Students were selected from urban, suburban, and rural school districts with approximately 1/3 of their sample coming from each district. Students indicated whether good grades, athletic ability, or popularity was most important to them. They also ranked four factors: grades, sports, looks, and money, in order of their importance for popularity. The questionnaire also asked for gender, grade level, and other demographic information.

Format

A data frame with 478 observations on the following 11 variables.

gender

a factor with levels boy girl

grade

grade in school

age

student age

race

a factor with levels other White

urban.rural

a factor with levels Rural Suburban Urban

school

a factor with levels Brentwood Elementary Brentwood Middle Brown Middle Elm Main Portage Ridge Sand Westdale Middle

goals

a factor with levels Grades Popular Sports

grades

rank of `make good grades' (1 = most important for popularity; 4 = least important)

sports

rank of `beging good at sports' (1 = most important for popularity; 4 = least important)

looks

rank of `beging handsome or pretty' (1 = most important for popularity; 4 = least important)

money

rank of `having lots of money' (1 = most important for popularity; 4 = least important)

Source

These data are available at DASL, the data and story library (https://dasl.datadescription.com/).

References

Chase, M. A., and Dummer, G. M. (1992), "The Role of Sports as a Social Determinant for Children," Research Quarterly for Exercise and Sport, 63, 418-424.

Examples

data(Kids)
#> Warning: data set ‘Kids’ not found
tally(goals ~ urban.rural, data = Kids)
#>          urban.rural
#> goals     Rural Suburban Urban
#>   Grades     57       87   103
#>   Popular    50       42    49
#>   Sports     42       22    26
chisq.test(tally(~ goals + urban.rural, data = Kids))
#> 
#> 	Pearson's Chi-squared test
#> 
#> data:  tally(~goals + urban.rural, data = Kids)
#> X-squared = 18.828, df = 4, p-value = 0.0008497
#>